+++ src/js/components/AppListComponent.jsx
@@ -5,6 +5,7 @@ var React = require("react/addons");
 var AppListViewTypes = require("../constants/AppListViewTypes");
 var AppStatus = require("../constants/AppStatus");
 var FilterTypes = require("../constants/FilterTypes");
+var HealthStatus = require("../constants/HealthStatus");
 var Messages = require("../constants/Messages");
 var States = require("../constants/States");
 var AppListItemComponent = require("./AppListItemComponent");
@@ -52,9 +53,9 @@ function getGroupHealth(groupHealth, appHealth) {
   });
 }
 
-function getInitialAppStatusesCount() {
-  return Object.values(AppStatus).reduce(function (memo, status) {
-    memo[status] = 0;
+function getInitialFilterCounts(object) {
+  return Object.values(object).reduce(function (memo, name) {
+    memo[name] = 0;
     return memo;
   }, {});
 }
@@ -171,11 +172,12 @@ var AppListComponent = React.createClass({
     });
   },
 
-  filterNodes: function (nodesSequence, appsStatusesCount) {
+  filterNodes: function (nodesSequence, filterCounts) {
     var props = this.props;
     var currentGroup = props.currentGroup;
     var filters = props.filters;
 
+    var filterHealth = filters[FilterTypes.HEALTH];
     var filterText = filters[FilterTypes.TEXT];
     var filterLabels = filters[FilterTypes.LABELS];
     var filterStatus = filters[FilterTypes.STATUS];
@@ -189,7 +191,10 @@ var AppListComponent = React.createClass({
     }
 
     nodesSequence.each(app => {
-      appsStatusesCount[app.status]++;
+      filterCounts.appsStatusesCount[app.status]++;
+      app.health.forEach(health => {
+        filterCounts.appsHealthCount[health.state] += health.quantity;
+      });
     });
 
     if (filterLabels != null && filterLabels.length > 0) {
@@ -217,16 +222,33 @@ var AppListComponent = React.createClass({
       });
     }
 
+    if (filterHealth != null && filterHealth.length > 0) {
+      nodesSequence = nodesSequence.filter(app => {
+        if (app.health == null) {
+          return false;
+        }
+
+        return filterHealth.some(healthFilter => {
+          return app.health.some(health =>
+            health.state === healthFilter && health.quantity > 0
+          );
+        });
+      });
+    }
+
     return nodesSequence;
   },
 
-  getGroupedNodes: function (apps, appsStatusesCount) {
+  getGroupedNodes: function (apps, filterCounts) {
     var currentGroup = this.props.currentGroup;
 
     var appsInGroup = apps.filter(app => app.id.startsWith(currentGroup));
 
     appsInGroup.forEach(app => {
-      appsStatusesCount[app.status]++;
+      filterCounts.appsStatusesCount[app.status]++;
+      app.health.forEach(health => {
+        filterCounts.appsHealthCount[health.state] += health.quantity;
+      });
     });
 
     return appsInGroup
@@ -264,19 +286,22 @@ var AppListComponent = React.createClass({
 
     var nodesSequence;
 
-    var appsStatusesCount = getInitialAppStatusesCount();
+    var filterCounts = {
+      appsStatusesCount: getInitialFilterCounts(AppStatus),
+      appsHealthCount: getInitialFilterCounts(HealthStatus)
+    };
 
     // Global search view - only display filtered apps
     if (this.hasFilters()) {
       appListViewType = AppListViewTypes.APP_LIST;
-      nodesSequence = this.filterNodes(lazy(state.apps), appsStatusesCount)
+      nodesSequence = this.filterNodes(lazy(state.apps), filterCounts)
         .sortBy((app) => {
           return app[sortKey];
         }, state.sortDescending);
 
     // Grouped node view
     } else {
-      nodesSequence = lazy(this.getGroupedNodes(state.apps, appsStatusesCount))
+      nodesSequence = lazy(this.getGroupedNodes(state.apps, filterCounts))
         // Alphabetically presort
         .sortBy((app) => {
           return app.id;
@@ -306,7 +331,7 @@ var AppListComponent = React.createClass({
       })
       .value();
 
-    AppsActions.emitAppStatusesCount(appsStatusesCount);
+    AppsActions.emitFilterCounts(filterCounts);
 
     return appListItems;
   },